+++ fabric/src/main/java/io/github/sakurawald/fuji/module/initializer/economy/gui/BalanceTopGui.java
@@ -52,7 +52,7 @@ public class BalanceTopGui extends PagedGui<GameProfileAndEconomyAccount> {
     protected void drawPagedGui() {
         super.drawPagedGui();
 
-        GuiHelper.PlayerSkull.fetchPlayerHeadTextures(this);
+        GuiHelper.PlayerSkull.fillPlayerHeadTextures(this);
     }
 
 }